refactor(namespace): reorganize file structure prior to creating build/publish process
