+++ src/common/i18n/consts.ts
@@ -1,5 +1,5 @@
-import { createNamespacePrepender } from '@namespace/createNamespacePrepender';
-import { Namespaced } from '@namespace/types';
+import { createNamespacePrepender } from "@namespace/createNamespacePrepender";
+import { Namespaced } from "@namespace/types";
 
 export const SUPPORTED_LOCALES = [
   'en',
