+++ src/common/namespace/src/isNamespaceTest.ts
