+++ src/common/namespace/src/isValidNamespace.ts
