+++ src/common/namespace/src/types/index.ts
