+++ src/common/namespace/src/types/namespace.ts
